config/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
chromadb==0.4.18
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
numpy>=1.21.0,<2.0.0
google-generativeai>=0.8.0
PyPDF2==3.0.1
python-multipart==0.0.6
redis==5.0.1
boto3==1.34.0
psutil==5.9.6
gunicorn==21.2.0
asyncio-mqtt==0.13.0
setuptools>=65.0.0
wheel>=0.37.0
SQLAlchemy
asyncpg
psycopg2-binary
aiosqlite
